Correção na medição do tempo e mudança da unidade para segundos

+++ src/arquivos.h
@@ -21,9 +21,9 @@ FILE *carrega_arq_leitura(char nome[]);
 FILE *cria_arq_escrita(char nome[]);
 FILE *carrega_arq_escrita(char nome[]);
 Login monta_login(char linha[]);
-LSE *lse_carrega(LSE *l, char nome_dados[], char nome_resultados[], long *tempo);
-void lse_avalia(LSE *l, char nome_testes[], char nome_resultados[], long *tempo);
-AVL *avl_carrega(AVL *t, char nome_dados[], char nome_resultados[], long *tempo);
-void avl_avalia(AVL *t, char nome_testes[], char nome_resultados[], long *tempo);
+LSE *lse_carrega(LSE *l, char nome_dados[], char nome_resultados[], float *tempo);
+void lse_avalia(LSE *l, char nome_testes[], char nome_resultados[], float *tempo);
+AVL *avl_carrega(AVL *t, char nome_dados[], char nome_resultados[], float *tempo);
+void avl_avalia(AVL *t, char nome_testes[], char nome_resultados[], float *tempo);
 
 #endif
